Extract observed source root placement resolver
